refactor(transition): change duration type to int
BREAKING CHANGE: TransitionController#setDuration now accepts int instead of long

+++ src/main/java/io/github/leawind/perspectiveapi/api/TransitionController.java
@@ -12,7 +12,7 @@ public interface TransitionController {
   boolean isInTransition(long now);
 
   /// Sets the transition duration in milliseconds.
-  void setDuration(long duration);
+  void setDuration(int duration);
 
   /// Sets the blending function used for easing.
   void setBlender(@NonNull Blender blender);
